Move impl Arbitrary for Instance to random/instance.rs
